Apply formatting, bug fixing, and improvements to the overall structure of sqrt
